+++ rake_helper/spec.rb
@@ -19,7 +19,7 @@ begin
 rescue LoadError
 end
 
-SPEC_PATTERN ='spec/**/*_spec.rb'
+SPEC_PATTERN ='spec/**/local_adapt_spec.rb'
 
 puts "Creating dummy libs"
 `gcc -r -c rake_helper/dummy.c -o rake_helper/dummy.a`